feat: adjustments to sales and customer actions
