+++ apps/shoetrak-app/app/_actions/Customers/index.ts
@@ -1,0 +1,1 @@
+export * from './GetRecentCustomers';
