+++ apps/shoetrak-app/app/_actions/Sales/index.ts
@@ -1,0 +1,1 @@
+export * from './GetSalesHistory';
