+++ apps/shoetrak-app/app/_actions/index.ts
@@ -1,2 +1,4 @@
 export * from './Dashboard/Dashboard';
+export * from './Customers';
+export * from './Sales';
 export * from './Session/Session';
